# docs/swagger.yaml
basePath: /
definitions:
...
  clips.ClipResponse:
    description: Complete information about an audio clip
    properties:
      auto_labeled:
        example: false
        type: boolean
      created_at:
        example: "2025-09-25T16:36:45Z"
        type: string
      duration:
        example: 15
        type: number
      error_message:
        example: 'failed to download source audio: HTTP 403'
        type: string
      extracted:
        example: true
        type: boolean
      filename:
        example: clip_052f3b9b-cc02-418c-a9ab-8f49534c01c8.wav
        type: string
      label:
        example: advertisement
        type: string
      label_confidence:
        example: 0.85
        type: number
      label_method:
        example: manual
        type: string
      original_end_time:
        example: 45
        type: number
      original_start_time:
        example: 30
        type: number
      podcast_index_episode_id:
        example: 12345
        type: integer
      size_bytes:
        example: 480078
        type: integer
      source_episode_url:
        example: https://example.com/episode.mp3
        type: string
      status:
        enum:
        - queued
        - processing
        - ready
        - failed
        example: ready
        type: string
      updated_at:
        example: "2025-09-25T16:36:47Z"
        type: string
      uuid:
        example: 052f3b9b-cc02-418c-a9ab-8f49534c01c8
        type: string
    type: object
  clips.CreateClipRequest:
    description: Request body for creating a new audio clip. Audio URL is automatically
      retrieved from cached episode data. If episode not cached, it will be fetched
      from Podcast Index API and synced to database.
    properties:
      end_time:
        example: 45
        type: number
      label:
        example: advertisement
        minLength: 1
        type: string
      podcast_index_episode_id:
        example: 12345
        minimum: 1
        type: integer
      start_time:
        example: 30
        minimum: 0
        type: number
    required:
    - end_time
    - label
    - podcast_index_episode_id
    type: object
  clips.UpdateLabelRequest:
    description: Request body for updating a clip's label
    properties:
      label:
        example: music
        minLength: 1
        type: string
    required:
    - label
    type: object
...
  episodes.CreateClipRequest:
    properties:
      end_time:
        example: 45
        type: number
      label:
        example: advertisement
        minLength: 1
        type: string
      start_time:
        example: 30
        minimum: 0
        type: number
    required:
    - end_time
    - label
    type: object
  episodes.EpisodeClipResponse:
    properties:
      approved:
        example: true
        type: boolean
      auto_labeled:
        example: false
        type: boolean
      created_at:
        example: "2025-10-02T13:00:00Z"
        type: string
      duration:
        example: 15
        type: number
      error_message:
        example: ""
        type: string
      extracted:
        example: false
        type: boolean
      filename:
        example: clip_a1b2c3d4.wav
        type: string
      label:
        example: advertisement
        type: string
      label_confidence:
        example: 0.95
        type: number
      label_method:
        enum:
        - manual
        - peak_detection
        example: manual
        type: string
      original_end_time:
        example: 45
        type: number
      original_start_time:
        example: 30
        type: number
      size_bytes:
        example: 480332
        type: integer
      status:
        enum:
        - detected
        - queued
        - processing
        - ready
        - failed
        example: queued
        type: string
      updated_at:
        example: "2025-10-02T13:00:00Z"
        type: string
      uuid:
        example: a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      details:
        description: Additional error details
      error:
        description: Error code/type
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@example.com
    name: Podcast Player API Support
    url: https://github.com/killallgit/killallplayer-api
  description: |-
    A REST API for podcast discovery and episode management.
    This API provides endpoints for searching podcasts, managing episodes,
    and generating waveforms for audio visualization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Podcast Player API
  version: "1.0"
paths:
...
  /api/v1/clips:
    get:
      description: |-
        Retrieve a paginated list of clips with optional filtering by label and processing status.
        Results are ordered by creation time (newest first). Use this endpoint to monitor clip processing
        or to browse available training data by label.
      parameters:
      - description: Filter clips by exact label match (e.g., 'advertisement')
        in: query
        name: label
        type: string
      - description: Filter by processing status
        enum:
        - queued
        - processing
        - ready
        - failed
        in: query
        name: status
        type: string
      - default: 100
        description: Maximum number of clips to return (1-1000)
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of clips to skip for pagination
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of clips matching the filters
          schema:
            items:
              $ref: '#/definitions/clips.ClipResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List clips with optional filtering
      tags:
      - clips
    post:
      consumes:
      - application/json
      description: |-
        Create a labeled audio segment from a podcast episode for machine learning training datasets.
        The clip is stored as metadata (time range + label) and will be extracted during dataset export.
        No audio processing occurs immediately - clips are materialized only when exporting the dataset.
        The exact time range specified is preserved (no padding or cropping to fixed duration).
      parameters:
      - description: Audio clip parameters with episode ID and time range in seconds
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clips.CreateClipRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Clip created successfully (status=pending, awaiting export)
          schema:
            $ref: '#/definitions/clips.ClipResponse'
        "400":
          description: Invalid request parameters (e.g., end_time <= start_time)
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error during clip creation
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create a new audio clip for ML training
      tags:
      - clips
  /api/v1/clips/{uuid}:
    delete:
      description: |-
        Permanently delete a clip from the database and remove its associated audio file from storage.
        This operation cannot be undone. If the clip is already deleted, returns success (idempotent).
      parameters:
      - description: Unique clip identifier (UUID format)
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: Clip deleted successfully (no content returned)
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error during deletion
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete a clip and its audio file
      tags:
      - clips
    get:
      description: |-
        Retrieve detailed information about a specific clip including its processing status,
        audio properties, and label. Check the 'status' field to determine if the clip is ready for use.
      parameters:
      - description: Unique clip identifier (UUID format)
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Clip details retrieved successfully
          schema:
            $ref: '#/definitions/clips.ClipResponse'
        "404":
          description: Clip with specified UUID not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get clip details by UUID
      tags:
      - clips
  /api/v1/clips/{uuid}/label:
    put:
      consumes:
      - application/json
      description: |-
        Change the label of an existing clip to reorganize training datasets.
        This operation moves the clip file to a new label directory in storage.
        Labels can be any string value for flexible categorization (e.g., "advertisement", "music", "speech").
      parameters:
      - description: Unique clip identifier (UUID format)
        in: path
        name: uuid
        required: true
        type: string
      - description: New label for categorization
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/clips.UpdateLabelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Label updated successfully
          schema:
            $ref: '#/definitions/clips.ClipResponse'
        "400":
          description: Invalid request (empty label or malformed JSON)
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Clip with specified UUID not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error or storage operation failed
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update a clip's label for re-categorization
      tags:
      - clips
  /api/v1/clips/export:
    get:
      description: |-
        Export all approved clips as a ZIP archive for machine learning training.
        Audio extraction happens on-demand during export - clips are materialized from their time ranges.
        The archive contains audio files organized by label directories and a JSONL manifest file
        with metadata for each clip. Audio files are in 16kHz mono WAV format with exact durations preserved.
        The manifest includes clip UUID, label, duration, source URL, and original time range for full traceability.
        Previously extracted clips are reused to avoid redundant processing.
      produces:
      - application/zip
      responses:
        "200":
          description: ZIP archive containing labeled audio clips and manifest.jsonl
          schema:
            type: file
        "500":
          description: Internal server error during export
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Export ML training dataset as ZIP
      tags:
      - clips
...
  /api/v1/episodes/{id}/clips:
    get:
      description: Get all clips created for this episode with optional status and
        approval filters
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by status
        enum:
        - queued
        - processing
        - ready
        - failed
        - detected
        in: query
        name: status
        type: string
      - description: Filter by approval status (true/false)
        in: query
        name: approved
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/episodes.EpisodeClipResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List clips for episode
      tags:
      - episodes
    post:
      consumes:
      - application/json
      description: Create a new audio clip from this episode at the specified time
        range. Manual clips are automatically approved and will be extracted during
        dataset export.
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip creation parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/episodes.CreateClipRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Clip created successfully (approved=true, status=pending)
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create clip for episode
      tags:
      - episodes
  /api/v1/episodes/{id}/clips/{uuid}:
    delete:
      description: Delete a clip and its audio file from this episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete clip
      tags:
      - episodes
    get:
      description: Get details of a specific clip for this episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Clip UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/episodes.EpisodeClipResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get clip details
      tags:
      - episodes
...
schemes:
- http
- https
swagger: "2.0"
...